+++ twitchbot/__init__.py
@@ -1,0 +1,5 @@
+from twitchbot import initialize
+from twitchbot import read
+from twitchbot import settings
+from twitchbot import socket
+from twitchbot import api
